Add the ControlsAccess trait
